(fix) reorganise project packages
