--- src/main/java/pl/glozaaleksandra/homebudget/CategoryRepositoryFactory.java
@@ -1,7 +1,0 @@
-package pl.glozaaleksandra.homebudget;
-
-import pl.glozaaleksandra.homebudget.category.CategoryRepository;
-
-public interface CategoryRepositoryFactory {
-  public CategoryRepository createCategoryRepository();
-}
