--- src/main/java/pl/glozaaleksandra/homebudget/CategoryRepositoryFactoryImpl.java
@@ -1,11 +1,0 @@
-package pl.glozaaleksandra.homebudget;
-
-import pl.glozaaleksandra.homebudget.category.CategoryRepository;
-import pl.glozaaleksandra.homebudget.category.ListBasedCategoryRepository;
-
-public class CategoryRepositoryFactoryImpl implements CategoryRepositoryFactory {
-  @Override
-  public CategoryRepository createCategoryRepository() {
-    return new ListBasedCategoryRepository();
-  }
-}
