+++ src/main/java/pl/glozaaleksandra/homebudget/controllers/category/CategoryControllery.java
@@ -1,9 +1,10 @@
-package pl.glozaaleksandra.homebudget;
+package pl.glozaaleksandra.homebudget.controllers.category;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import pl.glozaaleksandra.homebudget.category.Category;
+import pl.glozaaleksandra.homebudget.repository.category.Category;
+import pl.glozaaleksandra.homebudget.service.category.CategoryService;
 
 import java.util.List;
