+++ src/main/java/pl/glozaaleksandra/homebudget/service/category/CategoryService.java
@@ -1,9 +1,9 @@
-package pl.glozaaleksandra.homebudget;
+package pl.glozaaleksandra.homebudget.service.category;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import pl.glozaaleksandra.homebudget.category.Category;
-import pl.glozaaleksandra.homebudget.category.CategoryRepository;
+import pl.glozaaleksandra.homebudget.repository.category.Category;
+import pl.glozaaleksandra.homebudget.repository.category.CategoryRepository;
 
 import java.util.List;
